+++ src/models/pool.rs
@@ -13,9 +13,9 @@ pub struct Pool {
     pub owner: ObjectId, //Reference to address of account
     pub type_pool: TypePool,
     pub minting_fee: String,
-    pub begin_at: i32,
-    pub end_at: i32,
-    pub update_at: i32,
-    pub create_at: i32,
+    pub begin_at: i64,
+    pub end_at: i64,
+    pub update_at: i64,
+    pub create_at: i64,
 }
 pub const NAME: &str = "pool";
